requirements.txt
aiohttp>=3.9.0
websockets>=12.0
pyyaml>=6.0

# GPIO (Pi only — install lgpio via apt: sudo apt install python3-lgpio)
adafruit-circuitpython-dht>=4.0.0
RPLCD>=1.4.0
smbus2>=0.6.0
